add sweet alert

+++ src/Pages/MyToys/MyToys.jsx
@@ -1,12 +1,14 @@
 import { useContext, useEffect, useState } from "react";
 import { UserContext } from "../../Providers/Providers";
 import { FaEdit, FaTrashAlt } from "react-icons/fa";
-import { Link } from "react-router-dom";
+import Update from "../UpdateToy/Update";
+import Swal from "sweetalert2";
 
 
 const MyToys = () => {
     const { user } = useContext(UserContext);
     const [allToys, setAllToys] = useState([]);
+    const [toys, setToys] = useState();
 
     const email = user?.email;
     console.log(email)
@@ -18,21 +20,39 @@ const MyToys = () => {
     }, [url])
 
     const handleDeleteToy = (id) => {
-        fetch(`http://localhost:5000/cars/${id}`, {
-            method: 'DELETE'
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/cars/${id}`, {
+                    method: 'DELETE'
+                })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount > 0) {
+                        const remaining = allToys.filter(coffee => coffee._id !== id);
+                        setAllToys(remaining);
+                        Swal.fire(
+                            'Deleted!',
+                            'Your file has been deleted.',
+                            'success'
+                        )
+                    }
+                })            
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    const remaining = allToys.filter(coffee => coffee._id !== id);
-                    setAllToys(remaining);
-                    alert('delete successfully');
-                }
-            })
     }
 
     return (
         <div className="overflow-x-auto w-full">
+            <Update toys={toys}></Update>
             <table className="table w-full">
                 {/* head */}
                 <thead>
@@ -66,9 +86,9 @@ const MyToys = () => {
                             <td>{toy.quantity}</td>
                             <th>
                                 <div className="flex gap-5">
-                                    <Link to={`/update/${toy._id}`}>
-                                        <FaEdit></FaEdit>
-                                    </Link>
+                                    <label onClick={() => setToys(toy)} htmlFor="toy-details">
+                                        <FaEdit className="cursor-pointer"></FaEdit>
+                                    </label>
                                     <FaTrashAlt className="cursor-pointer" onClick={() => handleDeleteToy(toy._id)}></FaTrashAlt>
                                 </div>
                             </th>
